feat: add ReviewEventConsumer, TotalReview in hotel entity, replace BigDecimal to double in aggregate rating and add review serivce in docker-compose

+++ review-service/src/main/java/edu/hotel/review/dto/rating/RatingAggregateResponse.java
@@ -12,7 +12,7 @@ public class RatingAggregateResponse {
 
     private TargetType targetType;
 
-    private BigDecimal avgRating;
+    private Double avgRating;
 
     private Integer totalReviews;
